fix schedule err"No meeting with ID"
