+++ react/features/riff-platform/components/EditMeeting/index.js
@@ -2,18 +2,40 @@
 /* eslint-disable react/no-multi-comp */
 /* eslint-disable react/jsx-no-bind */
 
-import { Grid } from '@material-ui/core';
+import { Grid, Typography } from '@material-ui/core';
 import PropTypes from 'prop-types';
 import React, { useEffect } from 'react';
+import { useParams } from 'react-router-dom';
 
 import { connect } from '../../../base/redux';
-import { meetingReset } from '../../actions/meeting';
+import { getMeetingById, meetingReset } from '../../actions/meeting';
+import Loader from '../Loader';
 import SchedulerForm from '../Scheduler/SchedulerForm';
 import StyledPaper from '../StyledPaper';
 
+const errorMessage = err => (<Grid
+    container = { true }
+    item = { true }
+    justify = 'center'
+    xs = { 12 }><Typography color = 'error'>{err}</Typography></Grid>);
 
-const EditMeeting = ({ resetMeeting }) => {
+
+const EditMeeting = ({ resetMeeting, fetchMeetingById, meetingError, meetingLoading, meeting }) => {
     useEffect(() => () => resetMeeting(), []);
+
+    const { meetingId } = useParams();
+
+    useEffect(() => {
+        fetchMeetingById(meetingId);
+    }, [ meetingId ]);
+
+    if (meetingError) {
+        return errorMessage(meetingError);
+    }
+
+    if (meetingLoading) {
+        return <Loader />;
+    }
 
     return (
         <Grid
@@ -23,7 +45,9 @@ const EditMeeting = ({ resetMeeting }) => {
                 item = { true }
                 xs = { 12 }>
                 <StyledPaper title = 'Edit a meeting'>
-                    <SchedulerForm isEditing = { true } />
+                    <SchedulerForm
+                        isEditing = { true }
+                        meeting = { meeting } />
                 </StyledPaper>
             </Grid>
         </Grid>
@@ -31,14 +55,23 @@ const EditMeeting = ({ resetMeeting }) => {
 };
 
 EditMeeting.propTypes = {
+    fetchMeetingById: PropTypes.func,
+    meeting: PropTypes.object,
+    meetingError: PropTypes.string,
+    meetingLoading: PropTypes.bool,
     resetMeeting: PropTypes.func
 };
 
-const mapStateToProps = () => {
-    return { };
+const mapStateToProps = state => {
+    return {
+        meetingError: state['features/riff-platform'].meeting.error,
+        meetingLoading: state['features/riff-platform'].meeting.loading,
+        meeting: state['features/riff-platform'].meeting.meeting
+    };
 };
 const mapDispatchToProps = dispatch => {
     return {
+        fetchMeetingById: id => dispatch(getMeetingById(id)),
         resetMeeting: () => dispatch(meetingReset())
     };
 };
